refactor: rename image component
