+++ src/render/core/app/components/app-image/index.jsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import PulseLoader from 'components/util-pulse-loader'
-import gatewayHelper from 'helpers/gateway'
+import PulseLoader from 'components/util-pulse-loader/'
+import gatewayHelper from 'helpers/gateway/'
 import log from 'logger'
 
-export default class BoxImage extends React.PureComponent {
+export default class Image extends React.PureComponent {
 
     constructor(props) {
         super(props);
